# openapi.yaml
openapi: 3.0.3
info:
  title: Movie Suggestion API
  description: >
    Personalized movie recommendation API built in Go using clean architecture.
    Integrates with OMDB for movie data, Neo4j for graph-based recommendations,
    PostgreSQL for authentication, and AWS SQS for async movie imports.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create user
      description: Creates a new user. Requires `users:write` role.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Email already exists
          content:
            text/plain:
              schema:
                type: string
                example: email already exists
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Users]
      summary: List users
      description: >
        Returns a paginated list of users. Requires `users:read` role.

        **Visibility rule:**
        - Users with only `users:read` (without `users:write` or `*`) see only their own record.
          The `email` filter is automatically forced to the caller's email.
        - Users with `users:write` or `*` can list and filter all users.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          description: Filter by exact email address.
          schema:
            type: string
            format: email
            example: alice@example.com
        - name: name
          in: query
          description: Filter by name (case-insensitive partial match).
          schema:
            type: string
            example: Alice
        - name: page
          in: query
          description: Page number (1-based, default 1).
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Items per page (1–100, default 20).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListUsersResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user
      description: Returns the details of a user. Requires `users:read` role and ownership or wildcard (`*`).
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Users]
      summary: Patch user
      description: >
        Updates user fields (`name`, `password`, `roles`) with partial payload.
        Requires `users:write` role (or wildcard `*`).

        **Field ownership rules:**
        - Owner (token subject equals path `id`) can update `name`, `password`, and `roles`.
        - Non-owner can update only `roles`.
        - Non-owner attempting to update `name` or `password` receives `403`.

        **Roles rules:**
        - Replaces the entire roles list when provided.
        - Empty list is allowed.
        - Roles must belong to the fixed allowlist:
          `users:read`, `users:write`, `movies:read`, `movies-watch:write`, `movies:write`.

        **Password rule:**
        - Minimum 6 characters.
      operationId: patchUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatchUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/movies:
    get:
      tags: [Movies]
      summary: Get recommended movies
      description: >
        Returns personalized movie recommendations for the authenticated user
        using cursor-based pagination.
        The user ID is extracted from the token claims — no path parameter required.
        The algorithm is auto-selected based on the user's watch history
        or can be overridden via the `algorithm` query parameter.
        Title filtering is optional and applied as case-insensitive contains.
        Access is allowed for `movies:read`, `movies:write`, or wildcard `*`.

        **Auto-selection rules:**

        | Watch count | Algorithm |
        |-------------|-----------|
        | < 5 | POPULAR |
        | 5 – 19 | CONTENT_BASED |
        | ≥ 20 | COLLABORATIVE |
      operationId: getRecommendedMovies
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Number of recommended movies to return per page (default 10, max 50).
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: cursor
          in: query
          description: Opaque cursor token for pagination. Obtain from `nextCursor` or `prevCursor` in the previous response.
          schema:
            type: string
        - name: algorithm
          in: query
          description: Override the auto-selected recommendation algorithm.
          schema:
            $ref: "#/components/schemas/RecommendationAlgorithm"
        - name: title
          in: query
          description: Case-insensitive partial title filter applied together with algorithm results.
          schema:
            type: string
            example: matrix
      responses:
        "200":
          description: Paginated list of recommended movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsPageResponse"
        "400":
          description: Invalid limit or cursor parameter, or unknown algorithm
          content:
            text/plain:
              schema:
                type: string
                example: invalid cursor
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserID:
      name: id
      in: path
      required: true
      description: User UUID
      schema:
        type: string
        format: uuid
        example: 90b4fd08-ee46-4d54-a7c0-1bab4857cec3
...
  responses:
    BadRequest:
      description: Bad request
      content:
        text/plain:
          schema:
            type: string
            example: invalid request body
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        text/plain:
          schema:
            type: string
            example: forbidden
    NotFound:
      description: Resource not found
      content:
        text/plain:
          schema:
            type: string
            example: not found
    InternalError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: internal server error
  schemas:
...
    CreateUserRequest:
      type: object
      required: [name, email, password, roles]
      properties:
        name:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          example: s3cr3t
        roles:
          type: array
          items:
            type: string
          example: ["users:read", "users:write", "movies:read", "movie-watch:write"]
    CreateUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 90b4fd08-ee46-4d54-a7c0-1bab4857cec3
        name:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@example.com
        roles:
          type: array
          items:
            type: string
          example: ["users:read", "movies:read"]
        createdAt:
          type: string
          format: date-time
          example: "2026-05-30T12:00:00Z"
    PatchUserRequest:
      type: object
      description: Partial user update payload. At least one field must be provided.
      properties:
        name:
          type: string
          example: Alice Updated
        password:
          type: string
          format: password
          minLength: 6
          example: "123456"
        roles:
          type: array
          items:
            type: string
            enum:
              - users:read
              - users:write
              - movies:read
              - movies-watch:write
              - movies:write
          example: ["users:read", "movies:read"]
    PatchUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 90b4fd08-ee46-4d54-a7c0-1bab4857cec3
        name:
          type: string
          example: Alice Updated
        email:
          type: string
          format: email
          example: alice@example.com
        roles:
          type: array
          items:
            type: string
          example: ["users:read", "movies:read"]
        createdAt:
          type: string
          format: date-time
          example: "2026-05-30T12:00:00Z"
    User:
      type: object
      properties:
        ID:
          type: string
          format: uuid
          example: 90b4fd08-ee46-4d54-a7c0-1bab4857cec3
        Name:
          type: string
          example: William
        Email:
          type: string
          format: email
          example: william_cesar_santos@hotmail.com
        CreatedAt:
          type: string
          format: date-time
          example: "2026-05-30T12:00:00Z"
        CurrentAlgorithm:
          $ref: "#/components/schemas/RecommendationAlgorithm"
        WatchCount:
          type: integer
          example: 12
        LikeCount:
          type: integer
          example: 8
        DislikeCount:
          type: integer
          example: 2
...
    MovieSummary:
      type: object
      description: Reduced movie representation returned in list endpoints.
      properties:
        id:
          type: string
          format: uuid
          example: 4e5ebac2-d48d-4fc6-9319-dd52cf3be8b6
        title:
          type: string
          example: The Matrix
        year:
          type: string
          example: "1999"
        poster:
          type: string
          format: uri
          example: https://m.media-amazon.com/images/M/...
        imdbRating:
          type: number
          format: float
          example: 8.7
...
    ListUsersItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 90b4fd08-ee46-4d54-a7c0-1bab4857cec3
        name:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@example.com
        roles:
          type: array
          items:
            type: string
          example: ["users:read", "movies:read"]
        createdAt:
          type: string
          format: date-time
          example: "2026-05-30T12:00:00Z"
    ListUsersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ListUsersItem"
        total:
          type: integer
          description: Total number of matching users.
          example: 42
        page:
          type: integer
          description: Current page number (1-based).
          example: 1
        pageSize:
          type: integer
          description: Number of items per page.
          example: 20
    RecommendationsPageResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MovieSummary"
        nextCursor:
          type: string
          nullable: true
          description: Opaque cursor to retrieve the next page. Null when there is no next page.
          example: eyJvZmZzZXQiOjEwLCJ0b3RhbCI6NDB9...
        prevCursor:
          type: string
          nullable: true
          description: Opaque cursor to retrieve the previous page. Null when there is no previous page.
          example: eyJvZmZzZXQiOjAsInRvdGFsIjo0MH0=...
        hasNext:
          type: boolean
          description: Whether there is a next page.
          example: true
        hasPrev:
          type: boolean
          description: Whether there is a previous page.
          example: false
        limit:
          type: integer
          description: Page size used in this request.
          example: 10
        count:
          type: integer
          description: Number of items in this page.
          example: 10
        total:
          type: integer
          description: Total number of matching recommended movies.
          example: 40
    RecommendationAlgorithm:
      type: string
      enum:
        - POPULAR
        - CONTENT_BASED
        - COLLABORATIVE
        - HYBRID
        - SERENDIPITY
      example: CONTENT_BASED